requirements.txt
asgiref==3.8.1
dataclasses==0.6
Django==4.2.16
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djongo==1.3.7
dnspython==2.7.0
mongoengine==0.29.1
PyJWT==2.10.0
pymongo==3.11.4
pytz==2024.2
six==1.16.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
